feat: organised user functionality in its own app
