+++ recipes_app/recipes/forms.py
@@ -1,9 +1,1 @@
-from django.contrib.auth.forms import UserCreationForm
 
-from .models import Account
-
-
-class UserRegisterForm(UserCreationForm):
-    class Meta:
-        model = Account
-        fields = ['username', 'email', 'password1', 'password2']
