+++ recipes_app/recipes/templates/recipes/home.html
@@ -1,4 +1,4 @@
-{% extends 'recipes/base.html' %}
+{% extends 'base.html' %}
 
 
 {% block content_title %}
@@ -19,7 +19,7 @@
                      class="rounded-circle me-2">
                 <p class="card-text">Explore our recipes collection!</p>
 
-                <a href="{% url 'password_change' %}" class="link-primary">Change Password</a>
+                <a href="{% url 'user:password_change' %}" class="link-primary">Change Password</a>
 
 
             </div>
@@ -29,8 +29,8 @@
             <div class="card-body">
                 <h5 class="card-title">Hello!</h5>
                 <p class="card-text">You are not logged in.</p>
-                <p><a href="{% url 'login' %}?next={{ request.path }}" class="link-primary">Log in</a> or <a
-                        href="{% url 'recipes:signup' %}?next={{ request.path }}" class="link-primary">Sign up</a></p>
+                <p><a href="{% url 'user:login' %}?next={{ request.path }}" class="link-primary">Log in</a> or <a
+                        href="{% url 'user:signup' %}?next={{ request.path }}" class="link-primary">Sign up</a></p>
             </div>
         </div>
     {% endif %}
