+++ recipes_app/recipes/templates/recipes/profile.html
@@ -1,4 +1,4 @@
-{% extends 'recipes/base.html' %}
+{% extends 'base.html' %}
 {% load static %}
 
 {% block content_title %}
