+++ recipes_app/recipes/templates/recipes/recipe_detail.html
@@ -1,4 +1,4 @@
-{% extends 'recipes/base.html' %}
+{% extends 'base.html' %}
 {% load static %}
 
 {% block content_title %}
