+++ recipes_app/recipes/templates/recipes/recipe_list.html
@@ -1,4 +1,4 @@
-{% extends 'recipes/base.html' %}
+{% extends 'base.html' %}
 {% load static %}
 {% block content_title %}
     <div class="d-flex justify-content-between">
