--- recipes_app/recipes/templates/recipes/thank_you.html
@@ -1,5 +1,0 @@
-{% extends 'recipes/base.html' %}
-
-{% block content %}
-<h1>thank you</h1>
-{% endblock %}
