+++ recipes_app/templates/base.html
@@ -1,0 +1,120 @@
+{% load static %}
+
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport"
+          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <!-- CSS only -->
+
+    <link href="{% static 'recipes/bootstrap.css' %}" rel="stylesheet">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.2/font/bootstrap-icons.css">
+    <link href="https://unpkg.com/gridjs/dist/theme/mermaid.min.css" rel="stylesheet"/>
+
+    <link href="{% static 'recipes/style.css' %}" rel="stylesheet">
+
+    <title>Document</title>
+
+</head>
+
+<body>
+
+<main class="d-flex flex-nowrap">
+
+    <div class="d-flex flex-column flex-shrink-0 p-3 bg-light" style="width: 280px; height: 100vh">
+        <a href="{% url 'recipes:home' %}"
+           class="d-flex align-items-center mb-3 mb-md-0 me-md-auto link-dark text-decoration-none">
+            <i class="bi bi-bug-fill"></i>
+            <span class="ms-2 fs-4 me-2">Food</span>
+            <i class="bi bi-bug-fill"></i>
+        </a>
+        <hr>
+        <ul class="nav nav-pills flex-column mb-auto">
+            <li class="">
+                <a href="{% url 'recipes:home' %}" class="nav-link link-dark align-middle d-inline-flex">
+                    <i class="bi bi-house-door-fill"></i>
+                    <span class="ms-2 fs-4">Home</span>
+                </a>
+            </li>
+            {% if user.is_authenticated %}
+                <li>
+                    <a href="{% url 'recipes:list_recipe' %}" class="nav-link link-dark d-inline-flex">
+                        {#                    <img src="{% static 'recipes/cook-book.png' %}" width="30" height="30" alt="my_image">#}
+                        <i class="bi bi-journal-richtext"></i>
+                        <span class="ms-2 fs-4">Recipes</span>
+                    </a>
+                </li>
+                <li>
+                    <a href="{% url 'recipes:profile' %}" class="nav-link link-dark d-inline-flex">
+                        {#                    <img src="{% static 'recipes/cook-book.png' %}" width="30" height="30" alt="my_image">#}
+                        <i class="bi bi-person-lines-fill"></i>
+                        <span class="ms-2 fs-4">My Recipes</span>
+                    </a>
+                </li>
+            {% endif %}
+        </ul>
+        {% if user.is_authenticated %}
+            <hr>
+            <div class="dropdown">
+                <a href="#" class="d-flex align-items-center link-dark text-decoration-none dropdown-toggle"
+                   id="dropdownUser2" data-bs-toggle="dropdown" aria-expanded="false">
+                    <img src="/static/recipes/static_cdn/default_profile_image.png" alt="" width="32" height="32" class="rounded-circle me-2">
+                    <strong>{{ user.username }}</strong>
+                </a>
+                <ul class="dropdown-menu text-small shadow" aria-labelledby="dropdownUser2">
+                    <li><a class="dropdown-item" href="{% url 'recipes:add_recipe' %}">New Recipe</a></li>
+                    <li><a class="dropdown-item" href="#">Settings</a></li>
+                    <li><a class="dropdown-item" href="{% url 'recipes:profile' %}">Profile</a></li>
+                    <li>
+                        <hr class="dropdown-divider">
+                    </li>
+                    <li>
+                        {% if user.is_authenticated %}
+                            <a class="dropdown-item" href="{% url 'user:logout' %}">Log out</a>
+{#                            <a class="dropdown-item"><form method="POST"></form>Log out</a>#}
+
+                        {% endif %}
+                    </li>
+                </ul>
+            </div>
+        {% endif %}
+    </div>
+    <div class="d-flex flex-column w-100 pe-5 content">
+        <div class="mt-5 ms-5">
+            <div class="card">
+                <div class="card-body">
+                    {% block content_title %}
+
+                    {% endblock %}
+                </div>
+            </div>
+
+        </div>
+        <div class="card mt-5 ms-5">
+            <div class="card-body p-4">
+                {% if messages %}
+                    {% for message in messages %}
+                        <div {% if message.tags %} class="alert alert-{{ message.tags }}"{% endif %}>{{ message }}</div>
+                    {% endfor %}
+                {% endif %}
+                {% block content %}
+
+                {% endblock %}
+            </div>
+        </div>
+    </div>
+
+</main>
+
+
+<!-- JavaScript Bundle with Popper -->
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-pprn3073KE6tl6bjs2QrFaJGz5/SUsLqktiwsUTF55Jfv3qYSDhgCecCxMW52nD2"
+        crossorigin="anonymous"></script>
+{% block javascript %}
+
+{% endblock %}
+</body>
+</html>
