--- recipes_app/templates/registration/password_change_form.html
@@ -1,18 +1,0 @@
-{% extends 'recipes/base.html' %}
-{% load crispy_forms_tags %}
-
-{% block content_title %}
-    <h4>Change password</h4>
-{% endblock %}
-
-
-{% block content %}
-
-    <form method="POST">
-        {% csrf_token %}
-        {{ form|crispy }}
-
-        <button class="btn btn-outline-secondary" type="submit">Update</button>
-    </form>
-
-{% endblock %}
