+++ recipes_app/user/templates/user/password_change_done.html
@@ -1,4 +1,4 @@
-{% extends 'recipes/base.html' %}
+{% extends 'base.html' %}
 
 {% block content_title %}
     <h4>Done!</h4>
@@ -8,8 +8,7 @@
 
     <div class="container">
         <div class="d-flex flex-column">
-            <p class="m-auto">Your password has been reset. You may go ahead and <a href="{% url 'login' %}">log in</a>
-                now.</p>
+            <p class="m-auto">Your password has been updated.</p>
         </div>
     </div>
