+++ recipes_app/user/templates/user/password_change_form.html
@@ -1,15 +1,18 @@
-{% extends 'recipes/base.html' %}
-
+{% extends 'base.html' %}
 {% load crispy_forms_tags %}
 
 {% block content_title %}
-<h4>Sign up here</h4>
+    <h4>Change password</h4>
 {% endblock %}
 
+
 {% block content %}
+
     <form method="POST">
         {% csrf_token %}
-        {{ form | crispy }}
-        <input class="btn btn-outline-secondary" type="submit" value="Sign Up">
+        {{ form|crispy }}
+
+        <button class="btn btn-outline-secondary" type="submit">Update</button>
     </form>
+
 {% endblock %}
