+++ recipes_app/user/templates/user/password_reset_complete.html
@@ -1,4 +1,4 @@
-{% extends 'recipes/base.html' %}
+{% extends 'base.html' %}
 
 {% block content_title %}
     <h4>Done!</h4>
@@ -8,7 +8,8 @@
 
     <div class="container">
         <div class="d-flex flex-column">
-            <p class="m-auto">Your password has been updated.</p>
+            <p class="m-auto">Your password has been reset. You may go ahead and <a href="{% url 'user:login' %}">log in</a>
+                now.</p>
         </div>
     </div>
