+++ recipes_app/user/templates/user/password_reset_done.html
@@ -1,4 +1,4 @@
-{% extends 'recipes/base.html' %}
+{% extends 'base.html' %}
 
 {% block content_title %}
     <h4>Email sent!</h4>
