+++ recipes_app/user/templates/user/password_reset_email.html
@@ -2,7 +2,7 @@
 To initiate the password reset process for your Account {{ user.email }},
 click the link below:
 
-{{ protocol }}://{{ domain }}{% url 'password_reset_confirm' uidb64=uid token=token %}
+{{ protocol }}://{{ domain }}{% url 'user:password_reset_confirm' uidb64=uid token=token %}
 
 If clicking the link above doesn't work, please copy and paste the URL in a new browser
 window instead.
