+++ recipes_app/user/templates/user/password_reset_form.html
@@ -1,4 +1,4 @@
-{% extends 'recipes/base.html' %}
+{% extends 'base.html' %}
 {% load crispy_forms_tags %}
 
 {% block content_title %}
