+++ recipes_app/user/templates/user/signup.html
@@ -1,16 +1,15 @@
 {% extends 'base.html' %}
+
 {% load crispy_forms_tags %}
 
 {% block content_title %}
-    <h4>Add recipe</h4>
+<h4>Sign up here</h4>
 {% endblock %}
 
 {% block content %}
     <form method="POST">
         {% csrf_token %}
-
-        {{ form|crispy }}
-        <input class="btn btn-outline-secondary" type="submit" value="SUBMIT">
+        {{ form | crispy }}
+        <input class="btn btn-outline-secondary" type="submit" value="Sign Up">
     </form>
-
 {% endblock %}
